# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server ÖZINŞE.",
        "title": "ÖZINŞE API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
    "paths": {
...
        "/v1/users": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get Users List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/views.User"
                            }
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Get users",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user (Register)",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.ResponseId"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't hash password",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/users/profile": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.User"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Get user",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't update user data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.ResponseId"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't delete user",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.User"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't get user",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't update user data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.ResponseId"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't delete user",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "views.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "views.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "views.ResponseId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "views.Role": {
            "type": "object",
            "properties": {
                "age_categories": {
                    "type": "integer"
                },
                "genres": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "projects": {
                    "type": "integer"
                },
                "roles": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "types": {
                    "type": "integer"
                },
                "users": {
                    "type": "integer"
                }
            }
        },
...
        "views.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "description": "Id          int64     `json:\"id\"`",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "views.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "description": "Id          int64     `json:\"id\"`",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "views.User": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Role"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}